requirements.txt
flask
nltk
google-generativeai
markdown2
python_dotenv